refactor: 로그인 전용 레포 회원 레포로 이동
