+++ src/main/java/kr/inuappcenterportal/inuportal/domain/member/repository/SchoolLoginRepository.java
@@ -1,4 +1,4 @@
-package kr.inuappcenterportal.inuportal.global.oracleRepository;
+package kr.inuappcenterportal.inuportal.domain.member.repository;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
